chore: Fix CSSload for the fonts.
It now works, I had to go to a website to get the data URIs instead
of running openssl in bash (which is what I had done before). Need
to find a new way of doing this for the future.

ALSO: without caching this technique is pretty bad. There is a flash
of unstyled text on every page load (yuck!) MUST migrate to a proper
server so I can properly cache the fonts so that this doesn't happen.