fix: Remove h1 wrapping on multi-line title
This set up was wrong. It was causing strange behaviour by gulp-htmlmin
in some scenarios where it would move the position of the ending </h1>
tag to before the <hr> tag. Now we wrap in a div to get the same
effect more consistently.

Also, we removed the .--text class from the markup. We weren't using it
and it was spread out haphazardly.