docs: Update the resume about how to view the site
It no longer makes sense to host the page on gh-pages. There
are server side things that I need to do by going to a real
host and with the new architecture of having the index.html file
live in the dist folder instead of root it makes things too
annoying to deal with for too little pay off.